+++ include/CollisionObjective.h
@@ -25,7 +25,7 @@ class CollisionObjective : public ObjectiveFunction
 	SingleArmKinematicsSolver kSolver;
 	double norm_const = 41649.31279 * 10000;
 public:
-	CollisionObjective(int numOfJoints, double weight, P3D point, double radius, Robot* robot);
+	CollisionObjective(int numOfJoints, int weight, P3D point, double radius, Robot* robot);
 	~CollisionObjective();
 
 	virtual double computeValue(const dVector& p);
