+++ include/Simulation.h
@@ -59,7 +59,7 @@ public:
 	int IterationNum = 0;
 	MinimizerType MinimizerType;// = MinimizerType::GD;
 	void MakeStep();
-	AdvancedSimulation(VectorXd startPoint, VectorXd endPoint, int numOfPoints, std::vector<double> weights, 
+	AdvancedSimulation(VectorXd startPoint, VectorXd endPoint, int numOfPoints, std::vector<int> weights, 
 		int mt,Robot* robot, P3D finalCart, bool onlyFinalCart,std::vector<CollisionSphere> obstacles);
 	double ComputeValueAll();
 	double ComputeGradientAll();
