+++ src/StartFromFirstObjective.cpp
@@ -13,7 +13,7 @@ void printVector1(VectorXd pp)
 }
 
 
-StartFromFirstObjective::StartFromFirstObjective(const VectorXd & startPos, double weight)
+StartFromFirstObjective::StartFromFirstObjective(const VectorXd & startPos, int weight)
 {
 	m_startPos = startPos;
 	m_numOfJoints = m_startPos.rows();
